feat: extract local ext host and make it optional
