+++ rollup/rollup.config.ts
@@ -464,6 +464,7 @@ function resolveVscode (importee: string, importer?: string) {
 
 const input = {
   api: './src/api.ts',
+  localExtensionHost: './src/localExtensionHost.ts',
   extensions: './src/extensions.ts',
   services: './src/services.ts',
   l10n: './src/l10n.ts',
@@ -875,6 +876,10 @@ export default (args: Record<string, string>): rollup.RollupOptions[] => {
                 types: './services.d.ts',
                 default: './services.js'
               },
+              './localExtensionHost': {
+                types: './localExtensionHost.d.ts',
+                default: './localExtensionHost.js'
+              },
               './extensions': {
                 types: './extensions.d.ts',
                 default: './extensions.js'
@@ -1032,7 +1037,7 @@ export default (args: Record<string, string>): rollup.RollupOptions[] => {
                   const isVscodeFile = resolved.startsWith(VSCODE_SRC_DIST_DIR)
                   const isServiceOverride = path.dirname(resolved) === DIST_SERVICE_OVERRIDE_DIR_MAIN
                   const isNotExclusive = (isVscodeFile || isServiceOverride) && !exclusiveModules.has(resolvedWithExtension)
-                  const shouldBeShared = ['monaco.js', 'assets.js', 'lifecycle.js', 'workbench.js', 'missing-services.js'].includes(path.relative(DIST_DIR_MAIN, resolvedWithExtension))
+                  const shouldBeShared = ['extensions.js', 'monaco.js', 'assets.js', 'lifecycle.js', 'workbench.js', 'missing-services.js'].includes(path.relative(DIST_DIR_MAIN, resolvedWithExtension))
 
                   if (isNotExclusive || shouldBeShared) {
                     // Those modules will be imported from external monaco-vscode-api
